python/cerebras/synthetic/requirements.txt
fastapi 
uvicorn 
torch 
diffusers 
huggingface_hub 
safetensors
transformers
pillow
numpy
blenderproc
python-multipart